Move chooser forms into wagtail.admin.forms.chooser
